Created search_context/windows_test
