+++ test/unit/webdriver/search_context/android_test.py
@@ -16,6 +16,7 @@ import json
 
 import httpretty
 
+from appium.webdriver.webelement import WebElement as MobileWebElement
 from test.unit.helper.test_helper import (
     android_w3c_driver,
     appium_command,
@@ -71,6 +72,61 @@ class TestWebDriverAndroidSearchContext(object):
             body='{"value": []}'
         )
         els = driver.find_elements_by_android_data_matcher()
+
+        d = get_httpretty_request_body(httpretty.last_request())
+        assert d['using'] == '-android datamatcher'
+        assert d['value'] == '{}'
+        assert len(els) == 0
+
+    @httpretty.activate
+    def test_find_element_by_android_data_matcher(self):
+        driver = android_w3c_driver()
+        element = MobileWebElement(driver, 'element_id', w3c=True)
+        httpretty.register_uri(
+            httpretty.POST,
+            appium_command('/session/1234567890/element/element_id/element'),
+            body='{"value": {"element-6066-11e4-a52e-4f735466cecf": "child-element-id"}}'
+        )
+        el = element.find_element_by_android_data_matcher(
+            name='title', args=['title', 'Animation'], className='class name')
+
+        d = get_httpretty_request_body(httpretty.last_request())
+        assert d['using'] == '-android datamatcher'
+        value_dict = json.loads(d['value'])
+        assert value_dict['args'] == ['title', 'Animation']
+        assert value_dict['name'] == 'title'
+        assert value_dict['class'] == 'class name'
+        assert el.id == 'child-element-id'
+
+    @httpretty.activate
+    def test_find_elements_by_android_data_matcher(self):
+        driver = android_w3c_driver()
+        element = MobileWebElement(driver, 'element_id', w3c=True)
+        httpretty.register_uri(
+            httpretty.POST,
+            appium_command('/session/1234567890/element/element_id/elements'),
+            body='{"value": [{"element-6066-11e4-a52e-4f735466cecf": "child-element-id1"}, {"element-6066-11e4-a52e-4f735466cecf": "child-element-id2"}]}'
+        )
+        els = element.find_elements_by_android_data_matcher(name='title', args=['title', 'Animation'])
+
+        d = get_httpretty_request_body(httpretty.last_request())
+        assert d['using'] == '-android datamatcher'
+        value_dict = json.loads(d['value'])
+        assert value_dict['args'] == ['title', 'Animation']
+        assert value_dict['name'] == 'title'
+        assert els[0].id == 'child-element-id1'
+        assert els[1].id == 'child-element-id2'
+
+    @httpretty.activate
+    def test_find_elements_by_android_data_matcher_no_value(self):
+        driver = android_w3c_driver()
+        element = MobileWebElement(driver, 'element_id', w3c=True)
+        httpretty.register_uri(
+            httpretty.POST,
+            appium_command('/session/1234567890/element/element_id/elements'),
+            body='{"value": []}'
+        )
+        els = element.find_elements_by_android_data_matcher()
 
         d = get_httpretty_request_body(httpretty.last_request())
         assert d['using'] == '-android datamatcher'
